✨ feat(server): P2 消费端认证链路全量落地——REST 层(store/auth 5 端点 + store/account 6 端点,JWT 鉴权 extractor + 40100 短路体)、账户服务(资料/登录方式/换主邮箱/删号 v2.2 路由版)、i18n 错误消息四语言、ErrorCode::from_wire 数值映射;修 BIGINT UNSIGNED 列 i64 解码类型错配(u64);端到端全绿:OTP 发码→验证→建号→JWT 签发→带 token 访问 profile/identities/update→Redis 主存/集合索引全链验证通过

+++ server/crates/identity/src/service/auth.rs
@@ -303,8 +303,8 @@ pub async fn refresh(
         };
     }
     let row_data = RowData {
-        id: col!(i64, 0),
-        user_id: col!(i64, 1),
+        id: col!(u64, 0) as i64,
+        user_id: col!(u64, 1) as i64,
         old_access_jti: row
             .try_get_by_index::<Option<String>>(2)
             .ok()
